While processing patches, verify the contents of the lines they change
When a line is removed by a patch, or it specifies a context line, we
can check if the content of the line matches the content added to that
line by a previous patch. If not, then the patch series we are
processing is not consecutive and/or a patch would not apply cleanly
(at least not without offset).

Since we rely on line numbers, and not actual file contents, these
differences could cause false positives or negatives. This check
mitigates this problem somewhat, though it is still possible for
dependencies to  be missed when the patches are not a clean series.