parser: Handle chunks of length 1
In the unified diff format, when the length of a chunk is 1 (which can
only happen if there is no context, so when working with a file of
length 1), the length can be left out of the header.

Previously, the code would error out, since converting the None returned
by the regex to an integer is impossible. Now, this case is explicitely
handled.