Rewrite the ByLineFileAnalyzer
It now keeps two states during the analysis of each patch: The previous
state and the new state, moving lines from one to another. I previously
tried hard not to implement the analysis like this, because of the
overhead of keeping two states (which, in hindsight, is probably not so
bad at all...).

However, the previous implementation cannot be extended to support
analyzing git merge commits, which have multiple parents so multiple
previous states. The new approach can later be extended to support
merges.

In addition to changing the analysis method, the way proximity is
handled is also made a bit more consistent: Changes need to be twice the
proximity value apart to prevent a dependency. With the previous
implmentation, there seemed to be an inconsistency between adds and
deletes, two deletes needed the distance once, two adds or a delete and
an add needed the distance twice. This is now always the same.